remove unwanted classses in webapp-loader and appserver-utill modules
