--- modules/webapp-loader/src/main/java/org/wso2/appserver/webapp/loader/ClassLoaderEnvironment.java
@@ -1,67 +1,0 @@
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package org.wso2.appserver.webapp.loader;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-/**
- *  Store the environment configuration of the class loader.
- */
-public class ClassLoaderEnvironment {
-
-    private String environmentName;
-    private EnvironmentType environmentType;
-    private List<String> packageList = new ArrayList<>();
-    private List<String> resourcesList = new ArrayList<>();
-
-
-    public ClassLoaderEnvironment(String environmentName, EnvironmentType environmentType, List<String> packageList,
-                                  List<String> resourcesList) {
-
-        this.environmentName = environmentName;
-        this.environmentType = environmentType;
-
-        if (Optional.ofNullable(packageList).isPresent()) {
-            this.packageList = packageList;
-        }
-        if (Optional.ofNullable(resourcesList).isPresent()) {
-            this.resourcesList = resourcesList;
-        }
-
-    }
-
-    public List<String> getPackageList() {
-        return packageList;
-    }
-
-    public List<String> getResourcesList() {
-        return resourcesList;
-    }
-
-    public String getEnvironmentName() {
-        return environmentName;
-    }
-
-    public EnvironmentType getEnvironmentType() {
-        return environmentType;
-    }
-}
