--- modules/webapp-loader/src/main/java/org/wso2/appserver/webapp/loader/CompoundEnumeration.java
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package org.wso2.appserver.webapp.loader;
-
-import java.util.Arrays;
-import java.util.Enumeration;
-import java.util.NoSuchElementException;
-
-/**
- * @param <E> enumeration type
- */
-public class CompoundEnumeration<E> implements Enumeration<E> {
-    private Enumeration<E>[] enums;
-    private int index = 0;
-
-    public CompoundEnumeration(Enumeration<E>[] var1) {
-        this.enums = Arrays.copyOf(var1, var1.length);
-    }
-
-    private boolean next() {
-        while (this.index < this.enums.length) {
-            if (this.enums[this.index] != null && this.enums[this.index].hasMoreElements()) {
-                return true;
-            }
-
-            ++this.index;
-        }
-
-        return false;
-    }
-
-    public boolean hasMoreElements() {
-        return this.next();
-    }
-
-    public E nextElement() {
-        if (!this.next()) {
-            throw new NoSuchElementException();
-        } else {
-            return this.enums[this.index].nextElement();
-        }
-    }
-}
